requirements.txt
Django==1.6
djangorestframework==2.3.8